# api/openapi.yaml
openapi: 3.0.3
info:
  title: Companies API
  version: 1.0.0
  description: API for managing companies.
servers:
  - url: /
...
paths:
  /v1/companies:
    post:
      operationId: CreateCompany
      tags:
        - companies
      summary: Create a company.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Company created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/companies/{id}:
    parameters:
      - $ref: '#/components/parameters/CompanyID'
    get:
      operationId: GetCompany
      tags:
        - companies
      summary: Get a company.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Company found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: UpdateCompany
      tags:
        - companies
      summary: Partially update a company.
      description: Uses JSON Merge Patch semantics. Omitted fields remain unchanged and a null description clears the description.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: Company updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: DeleteCompany
      tags:
        - companies
      summary: Delete a company.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Company deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    CompanyID:
      name: id
      in: path
      required: true
      description: Company identifier.
      schema:
        type: string
  schemas:
    Company:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
        - employees_count
        - registered
        - type
        - created_at
        - updated_at
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 15
        description:
          type: string
          maxLength: 3000
          nullable: true
        employees_count:
          type: integer
          format: int32
          minimum: 0
        registered:
          type: boolean
        type:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    CreateCompanyRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - employees_count
        - registered
        - type
      properties:
        name:
          type: string
          description: Required.
          minLength: 1
          maxLength: 15
        description:
          type: string
          maxLength: 3000
          nullable: true
        employees_count:
          type: integer
          format: int32
          minimum: 0
          description: Required.
        registered:
          type: boolean
          description: Required.
        type:
          type: string
          description: Required.
    UpdateCompanyRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 15
        description:
          type: string
          maxLength: 3000
          nullable: true
        employees_count:
          type: integer
          format: int32
          minimum: 0
        registered:
          type: boolean
        type:
          type: string
    Error:
      type: object
      additionalProperties: false
      required:
        - status
        - title
      properties:
        status:
          type: integer
          format: int32
        title:
          type: string
        detail:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Violation:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Company was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Company name is already taken.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT