# internal/port/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package http

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/nullable"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Company defines model for Company.
type Company struct {
	CreatedAt      *time.Time                `json:"created_at,omitempty"`
	Description    nullable.Nullable[string] `json:"description"`
	EmployeesCount int32                     `json:"employees_count"`
	Id             *string                   `json:"id,omitempty"`
	Name           string                    `json:"name"`
	Registered     bool                      `json:"registered"`
	Type           string                    `json:"type"`
	UpdatedAt      *time.Time                `json:"updated_at,omitempty"`
}

// CreateCompanyRequest defines model for CreateCompanyRequest.
type CreateCompanyRequest struct {
	Description nullable.Nullable[string] `json:"description,omitempty"`

	// EmployeesCount Required.
	EmployeesCount int32 `json:"employees_count"`

	// Name Required.
	Name string `json:"name"`

	// Registered Required.
	Registered bool `json:"registered"`

	// Type Required.
	Type string `json:"type"`
}

// Error defines model for Error.
type Error struct {
	Detail     *string      `json:"detail,omitempty"`
	Status     int32        `json:"status"`
	Title      string       `json:"title"`
	Violations *[]Violation `json:"violations,omitempty"`
}

// UpdateCompanyRequest defines model for UpdateCompanyRequest.
type UpdateCompanyRequest struct {
	Description    nullable.Nullable[string] `json:"description,omitempty"`
	EmployeesCount *int32                    `json:"employees_count,omitempty"`
	Name           *string                   `json:"name,omitempty"`
	Registered     *bool                     `json:"registered,omitempty"`
	Type           *string                   `json:"type,omitempty"`
}

// Violation defines model for Violation.
type Violation struct {
	Field   string `json:"field"`
	Message string `json:"message"`
}

// CompanyID defines model for CompanyID.
type CompanyID = string

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Conflict defines model for Conflict.
type Conflict = Error

// NotFound defines model for NotFound.
type NotFound = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// CreateCompanyJSONRequestBody defines body for CreateCompany for application/json ContentType.
type CreateCompanyJSONRequestBody = CreateCompanyRequest

// UpdateCompanyJSONRequestBody defines body for UpdateCompany for application/json ContentType.
type UpdateCompanyJSONRequestBody = UpdateCompanyRequest

// UpdateCompanyApplicationMergePatchPlusJSONRequestBody defines body for UpdateCompany for application/merge-patch+json ContentType.
type UpdateCompanyApplicationMergePatchPlusJSONRequestBody = UpdateCompanyRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create a company.
	// (POST /v1/companies)
	CreateCompany(w http.ResponseWriter, r *http.Request)
	// Delete a company.
	// (DELETE /v1/companies/{id})
	DeleteCompany(w http.ResponseWriter, r *http.Request, id CompanyID)
	// Get a company.
	// (GET /v1/companies/{id})
	GetCompany(w http.ResponseWriter, r *http.Request, id CompanyID)
	// Partially update a company.
	// (PATCH /v1/companies/{id})
	UpdateCompany(w http.ResponseWriter, r *http.Request, id CompanyID)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Create a company.
// (POST /v1/companies)
func (_ Unimplemented) CreateCompany(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete a company.
// (DELETE /v1/companies/{id})
func (_ Unimplemented) DeleteCompany(w http.ResponseWriter, r *http.Request, id CompanyID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get a company.
// (GET /v1/companies/{id})
func (_ Unimplemented) GetCompany(w http.ResponseWriter, r *http.Request, id CompanyID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Partially update a company.
// (PATCH /v1/companies/{id})
func (_ Unimplemented) UpdateCompany(w http.ResponseWriter, r *http.Request, id CompanyID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CreateCompany operation middleware
func (siw *ServerInterfaceWrapper) CreateCompany(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateCompany(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteCompany operation middleware
func (siw *ServerInterfaceWrapper) DeleteCompany(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id CompanyID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteCompany(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetCompany operation middleware
func (siw *ServerInterfaceWrapper) GetCompany(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id CompanyID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetCompany(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateCompany operation middleware
func (siw *ServerInterfaceWrapper) UpdateCompany(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id CompanyID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateCompany(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/companies", wrapper.CreateCompany)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/v1/companies/{id}", wrapper.DeleteCompany)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/companies/{id}", wrapper.GetCompany)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/v1/companies/{id}", wrapper.UpdateCompany)
	})

	return r
}

type BadRequestJSONResponse Error

type ConflictJSONResponse Error

type NotFoundJSONResponse Error

type UnauthorizedJSONResponse Error

type CreateCompanyRequestObject struct {
	Body *CreateCompanyJSONRequestBody
}

type CreateCompanyResponseObject interface {
	VisitCreateCompanyResponse(w http.ResponseWriter) error
}

type CreateCompany201JSONResponse Company

func (response CreateCompany201JSONResponse) VisitCreateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateCompany400JSONResponse struct{ BadRequestJSONResponse }

func (response CreateCompany400JSONResponse) VisitCreateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateCompany401JSONResponse struct{ UnauthorizedJSONResponse }

func (response CreateCompany401JSONResponse) VisitCreateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type CreateCompany409JSONResponse struct{ ConflictJSONResponse }

func (response CreateCompany409JSONResponse) VisitCreateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type DeleteCompanyRequestObject struct {
	Id CompanyID `json:"id"`
}

type DeleteCompanyResponseObject interface {
	VisitDeleteCompanyResponse(w http.ResponseWriter) error
}

type DeleteCompany204Response struct {
}

func (response DeleteCompany204Response) VisitDeleteCompanyResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteCompany400JSONResponse struct{ BadRequestJSONResponse }

func (response DeleteCompany400JSONResponse) VisitDeleteCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type DeleteCompany401JSONResponse struct{ UnauthorizedJSONResponse }

func (response DeleteCompany401JSONResponse) VisitDeleteCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type DeleteCompany404JSONResponse struct{ NotFoundJSONResponse }

func (response DeleteCompany404JSONResponse) VisitDeleteCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetCompanyRequestObject struct {
	Id CompanyID `json:"id"`
}

type GetCompanyResponseObject interface {
	VisitGetCompanyResponse(w http.ResponseWriter) error
}

type GetCompany200JSONResponse Company

func (response GetCompany200JSONResponse) VisitGetCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetCompany400JSONResponse struct{ BadRequestJSONResponse }

func (response GetCompany400JSONResponse) VisitGetCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetCompany401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetCompany401JSONResponse) VisitGetCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type GetCompany404JSONResponse struct{ NotFoundJSONResponse }

func (response GetCompany404JSONResponse) VisitGetCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type UpdateCompanyRequestObject struct {
	Id                                CompanyID `json:"id"`
	JSONBody                          *UpdateCompanyJSONRequestBody
	ApplicationMergePatchPlusJSONBody *UpdateCompanyApplicationMergePatchPlusJSONRequestBody
}

type UpdateCompanyResponseObject interface {
	VisitUpdateCompanyResponse(w http.ResponseWriter) error
}

type UpdateCompany200JSONResponse Company

func (response UpdateCompany200JSONResponse) VisitUpdateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type UpdateCompany400JSONResponse struct{ BadRequestJSONResponse }

func (response UpdateCompany400JSONResponse) VisitUpdateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type UpdateCompany401JSONResponse struct{ UnauthorizedJSONResponse }

func (response UpdateCompany401JSONResponse) VisitUpdateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type UpdateCompany404JSONResponse struct{ NotFoundJSONResponse }

func (response UpdateCompany404JSONResponse) VisitUpdateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type UpdateCompany409JSONResponse struct{ ConflictJSONResponse }

func (response UpdateCompany409JSONResponse) VisitUpdateCompanyResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create a company.
	// (POST /v1/companies)
	CreateCompany(ctx context.Context, request CreateCompanyRequestObject) (CreateCompanyResponseObject, error)
	// Delete a company.
	// (DELETE /v1/companies/{id})
	DeleteCompany(ctx context.Context, request DeleteCompanyRequestObject) (DeleteCompanyResponseObject, error)
	// Get a company.
	// (GET /v1/companies/{id})
	GetCompany(ctx context.Context, request GetCompanyRequestObject) (GetCompanyResponseObject, error)
	// Partially update a company.
	// (PATCH /v1/companies/{id})
	UpdateCompany(ctx context.Context, request UpdateCompanyRequestObject) (UpdateCompanyResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// CreateCompany operation middleware
func (sh *strictHandler) CreateCompany(w http.ResponseWriter, r *http.Request) {
	var request CreateCompanyRequestObject

	var body CreateCompanyJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateCompany(ctx, request.(CreateCompanyRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateCompany")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateCompanyResponseObject); ok {
		if err := validResponse.VisitCreateCompanyResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// DeleteCompany operation middleware
func (sh *strictHandler) DeleteCompany(w http.ResponseWriter, r *http.Request, id CompanyID) {
	var request DeleteCompanyRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteCompany(ctx, request.(DeleteCompanyRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteCompany")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeleteCompanyResponseObject); ok {
		if err := validResponse.VisitDeleteCompanyResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetCompany operation middleware
func (sh *strictHandler) GetCompany(w http.ResponseWriter, r *http.Request, id CompanyID) {
	var request GetCompanyRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetCompany(ctx, request.(GetCompanyRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetCompany")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetCompanyResponseObject); ok {
		if err := validResponse.VisitGetCompanyResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// UpdateCompany operation middleware
func (sh *strictHandler) UpdateCompany(w http.ResponseWriter, r *http.Request, id CompanyID) {
	var request UpdateCompanyRequestObject

	request.Id = id
	if strings.HasPrefix(r.Header.Get("Content-Type"), "application/json") {

		var body UpdateCompanyJSONRequestBody
		if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
			sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
			return
		}
		request.JSONBody = &body
	}
	if strings.HasPrefix(r.Header.Get("Content-Type"), "application/merge-patch+json") {

		var body UpdateCompanyApplicationMergePatchPlusJSONRequestBody
		if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
			sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
			return
		}
		request.ApplicationMergePatchPlusJSONBody = &body
	}

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.UpdateCompany(ctx, request.(UpdateCompanyRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpdateCompany")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(UpdateCompanyResponseObject); ok {
		if err := validResponse.VisitUpdateCompanyResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}